Extract custom task as first level class
